battle frontier mon ability set option

+++ include/data.h
@@ -56,7 +56,7 @@ struct TrainerMon
     u16 moves[4];
     u16 species;
     u16 heldItem;
-    u16 ability;
+    u8 ability;
     u8 lvl;
     u8 ball;
     u8 friendship;
